refactor: Split Navigator into `.cpp` file
