+++ Core/src/Navigation/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(
     ActsCore
     PRIVATE
+        Navigator.cpp
         NavigationStream.cpp
         TryAllNavigationPolicy.cpp
         SurfaceArrayNavigationPolicy.cpp
